+++ doc/version_history.rst
@@ -12,6 +12,7 @@ v1.19.0
 * Add the `rules.MTM2OutClosedLoopControl` rule.
 * Add the `rules.MTM2TotalForceMoment` rule.
 * Add the `rules.MTM2ForceError` rule.
+* Add the `rules.MTM2MirrorTemperature` rule.
 
 v1.18.1
 -------